+++ examples/resources/pingfederate_idp_sts_request_parameters_contract/resource.tf
@@ -1,0 +1,9 @@
+resource "pingfederate_idp_sts_request_parameters_contract" "requestParametersContract" {
+  contract_id = "mycontract"
+  name        = "My Contract"
+  parameters = [
+    "firstparam",
+    "secondparam",
+    "thirdparam"
+  ]
+}